Split angle workspace UI components
