+++ frontend/src/components/tools/AngleWorkspace.tsx
@@ -4,7 +4,6 @@
 
 import deepmerge from 'deepmerge';
 import clsx from 'clsx';
-import dynamic from 'next/dynamic';
 import Link from 'next/link';
 import { usePathname, useSearchParams } from 'next/navigation';
 import useSWR from 'swr';
@@ -17,8 +16,6 @@ import {
   type ChangeEvent,
   type ClipboardEvent as ReactClipboardEvent,
   type DragEvent as ReactDragEvent,
-  type MouseEvent as ReactMouseEvent,
-  type PointerEvent as ReactPointerEvent,
 } from 'react';
 import { ArrowLeft, Download, Images, Loader2, Plus, Upload, WandSparkles, X } from 'lucide-react';
 import { HeaderBar } from '@/components/HeaderBar';
@@ -33,6 +30,9 @@ import { resolveAngleEngineForParams } from '@/lib/tools-angle';
 import { useRequireAuth } from '@/hooks/useRequireAuth';
 import { FEATURES } from '@/content/feature-flags';
 import type { AngleToolEngineId, AngleToolNumericParams, AngleToolResponse } from '@/types/tools-angle';
+import { AngleImageLibraryModal } from './angle/_components/angle-image-library-modal';
+import { AngleOrbitSelector } from './angle/_components/angle-orbit-selector';
+import { AngleOutputMosaic } from './angle/_components/angle-output-mosaic';
 import { DEFAULT_ANGLE_COPY, type AngleCopy } from './angle/_lib/angle-workspace-copy';
 import {
   ANGLE_GUEST_EXAMPLE_OUTPUT_URL,
@@ -51,304 +51,12 @@ import {
   uploadImage,
 } from './angle/_lib/angle-workspace-helpers';
 import type {
-  AnglePreviewImage,
   BillingProductResponse,
   LibraryAsset,
-  LibraryAssetsResponse,
   PersistedAngleToolState,
   RecentAngleJobEntry,
   UploadedImage,
 } from './angle/_lib/angle-workspace-types';
-
-const AngleOrbitCanvas = dynamic(() => import('@/components/tools/AngleOrbitCanvas'), {
-  ssr: false,
-  loading: () => <div className="h-full w-full bg-[#f6f8fc]" aria-hidden="true" />,
-});
-
-function AngleOutputMosaic({
-  outputs,
-  selectedIndex,
-  onSelect,
-  onDownload,
-  onAddToLibrary,
-  libraryDisabled = false,
-  compact = false,
-  singleRow = false,
-  copy,
-}: {
-  outputs: AnglePreviewImage[];
-  selectedIndex?: number;
-  onSelect?: (index: number) => void;
-  onDownload?: (url: string) => void;
-  onAddToLibrary?: (url: string) => void;
-  libraryDisabled?: boolean;
-  compact?: boolean;
-  singleRow?: boolean;
-  copy: AngleCopy;
-}) {
-  if (!outputs.length) return null;
-
-  if (outputs.length === 1) {
-    const output = outputs[0];
-    return (
-      <div className={clsx('overflow-hidden rounded-card border border-border bg-bg', compact ? 'h-full' : '')}>
-        <div className="relative">
-          <img
-            src={output.thumbUrl ?? output.url}
-            alt=""
-            className={clsx('w-full', compact ? 'h-full object-cover' : 'h-[360px] object-contain')}
-          />
-          {!compact && (onDownload || onAddToLibrary) ? (
-            <div className="absolute bottom-3 right-3 flex items-center gap-2">
-              {onAddToLibrary ? (
-                <button
-                  type="button"
-                  title={libraryDisabled ? copy.alreadyInLibrary : copy.addToLibrary}
-                  onClick={() => onAddToLibrary(output.url)}
-                  disabled={libraryDisabled}
-                  className="inline-flex h-8 w-8 items-center justify-center rounded-full border border-surface-on-media-25 bg-surface-on-media-dark-55 text-on-inverse transition hover:bg-surface-on-media-dark-70 disabled:cursor-default disabled:opacity-50"
-                >
-                  <Plus className="h-4 w-4" />
-                </button>
-              ) : null}
-              {onDownload ? (
-                <button
-                  type="button"
-                  title={copy.download}
-                  onClick={() => onDownload(output.url)}
-                  className="inline-flex h-8 w-8 items-center justify-center rounded-full border border-surface-on-media-25 bg-surface-on-media-dark-55 text-on-inverse transition hover:bg-surface-on-media-dark-70"
-                >
-                  <Download className="h-4 w-4" />
-                </button>
-              ) : null}
-            </div>
-          ) : null}
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className={clsx(singleRow ? 'grid grid-cols-4 gap-2' : 'grid grid-cols-2 gap-3')}>
-      {outputs.slice(0, 4).map((output, index) => {
-        const selected = index === selectedIndex;
-        const content = (
-          <>
-            <div className={clsx('relative overflow-hidden bg-bg', singleRow ? 'aspect-[4/5]' : 'aspect-square')}>
-              <img src={output.thumbUrl ?? output.url} alt="" className="h-full w-full object-cover" />
-              {!compact && selected && (onDownload || onAddToLibrary) ? (
-                <div className="absolute bottom-2 right-2 flex items-center gap-2">
-                  {onAddToLibrary ? (
-                    <button
-                      type="button"
-                      title={libraryDisabled ? copy.alreadyInLibrary : copy.addToLibrary}
-                      onClick={(event) => {
-                        event.stopPropagation();
-                        onAddToLibrary(output.url);
-                      }}
-                      disabled={libraryDisabled}
-                      className="inline-flex h-7 w-7 items-center justify-center rounded-full border border-surface-on-media-25 bg-surface-on-media-dark-55 text-on-inverse transition hover:bg-surface-on-media-dark-70 disabled:cursor-default disabled:opacity-50"
-                    >
-                      <Plus className="h-3.5 w-3.5" />
-                    </button>
-                  ) : null}
-                  {onDownload ? (
-                    <button
-                      type="button"
-                      title={copy.download}
-                      onClick={(event) => {
-                        event.stopPropagation();
-                        onDownload(output.url);
-                      }}
-                      className="inline-flex h-7 w-7 items-center justify-center rounded-full border border-surface-on-media-25 bg-surface-on-media-dark-55 text-on-inverse transition hover:bg-surface-on-media-dark-70"
-                    >
-                      <Download className="h-3.5 w-3.5" />
-                    </button>
-                  ) : null}
-                </div>
-              ) : null}
-            </div>
-            {!compact && !singleRow ? (
-              <div className="px-2 py-1 text-left text-[11px] text-text-secondary">
-                {copy.angleLabel.replace('{index}', String(index + 1))}
-                {selected ? ` · ${copy.angleSelected}` : ''}
-              </div>
-            ) : null}
-          </>
-        );
-
-        if (!onSelect) {
-          return (
-            <div key={`${output.url}-${index}`} className="overflow-hidden rounded-card border border-border bg-bg">
-              {content}
-            </div>
-          );
-        }
-
-        return (
-          <button
-            key={`${output.url}-${index}`}
-            type="button"
-            onClick={() => onSelect(index)}
-            className={clsx(
-              'overflow-hidden rounded-card border bg-bg text-left transition',
-              selected ? 'border-brand ring-1 ring-brand/40' : 'border-border hover:border-brand/40'
-            )}
-          >
-            {content}
-          </button>
-        );
-      })}
-    </div>
-  );
-}
-
-function AngleOrbitSelector({
-  params,
-  onParamsChange,
-  generateBestAngles,
-  onGenerateBestAnglesChange,
-  supportsMultiOutput,
-  sourceImage,
-  copy,
-}: {
-  params: AngleToolNumericParams;
-  onParamsChange: (next: AngleToolNumericParams) => void;
-  generateBestAngles: boolean;
-  onGenerateBestAnglesChange: (value: boolean) => void;
-  supportsMultiOutput: boolean;
-  sourceImage?: UploadedImage | null;
-  copy: AngleCopy;
-}) {
-  const dragStateRef = useRef<{
-    startX: number;
-    startY: number;
-    startRotation: number;
-    startTilt: number;
-  } | null>(null);
-  const [dragging, setDragging] = useState(false);
-
-  const handlePointerDown = (event: ReactPointerEvent<HTMLDivElement>) => {
-    event.currentTarget.setPointerCapture(event.pointerId);
-    dragStateRef.current = {
-      startX: event.clientX,
-      startY: event.clientY,
-      startRotation: params.rotation,
-      startTilt: params.tilt,
-    };
-    setDragging(true);
-  };
-
-  const handlePointerMove = (event: ReactPointerEvent<HTMLDivElement>) => {
-    if (!dragStateRef.current) return;
-    const dx = event.clientX - dragStateRef.current.startX;
-    const dy = event.clientY - dragStateRef.current.startY;
-    onParamsChange(
-      sanitizeParams({
-        rotation: dragStateRef.current.startRotation - dx * 0.55,
-        tilt: dragStateRef.current.startTilt + dy * 0.18,
-        zoom: params.zoom,
-      })
-    );
-  };
-
-  const handlePointerUp = () => {
-    dragStateRef.current = null;
-    setDragging(false);
-  };
-
-  const nudge = (deltaRotation: number, deltaTilt: number) => {
-    onParamsChange(
-      sanitizeParams({
-        rotation: params.rotation + deltaRotation,
-        tilt: params.tilt + deltaTilt,
-        zoom: params.zoom,
-      })
-    );
-  };
-
-  return (
-    <div className="rounded-card border border-border bg-bg p-4">
-      <div className="mb-2 text-center">
-        <p className="text-xs font-semibold uppercase tracking-micro text-text-muted">{copy.orbitTitle}</p>
-        <p className="mt-1 text-xs text-text-secondary">{copy.orbitHint}</p>
-      </div>
-
-      <div
-        onPointerDown={handlePointerDown}
-        onPointerMove={handlePointerMove}
-        onPointerUp={handlePointerUp}
-        onPointerCancel={handlePointerUp}
-        className={clsx(
-          'relative mx-auto w-full max-w-[360px] touch-none select-none rounded-card border border-border/80 bg-[#f6f8fc] p-3',
-          dragging ? 'cursor-grabbing' : 'cursor-grab'
-        )}
-      >
-        <div className="pointer-events-none relative mx-auto h-[320px] w-[320px] overflow-hidden rounded-card">
-          <AngleOrbitCanvas params={params} sourceImage={sourceImage} generateBestAngles={generateBestAngles} />
-
-          <button
-            type="button"
-            className="pointer-events-auto absolute left-1/2 top-2 -translate-x-1/2 rounded-full border border-border/80 bg-surface/80 px-2 py-1 text-xs text-text-secondary hover:bg-surface"
-            onClick={() => nudge(0, -4)}
-          >
-            ▲
-          </button>
-          <button
-            type="button"
-            className="pointer-events-auto absolute left-2 top-1/2 -translate-y-1/2 rounded-full border border-border/80 bg-surface/80 px-2 py-1 text-xs text-text-secondary hover:bg-surface"
-            onClick={() => nudge(12, 0)}
-          >
-            ◀
-          </button>
-          <button
-            type="button"
-            className="pointer-events-auto absolute right-2 top-1/2 -translate-y-1/2 rounded-full border border-border/80 bg-surface/80 px-2 py-1 text-xs text-text-secondary hover:bg-surface"
-            onClick={() => nudge(-12, 0)}
-          >
-            ▶
-          </button>
-          <button
-            type="button"
-            className="pointer-events-auto absolute bottom-2 left-1/2 -translate-x-1/2 rounded-full border border-border/80 bg-surface/80 px-2 py-1 text-xs text-text-secondary hover:bg-surface"
-            onClick={() => nudge(0, 4)}
-          >
-            ▼
-          </button>
-        </div>
-      </div>
-
-      <div className="mt-3 flex items-center justify-between text-xs text-text-secondary">
-        <span>{copy.rotationShort}: {Math.round(params.rotation)}°</span>
-        <span>{copy.tiltShort}: {Math.round(params.tilt)}°</span>
-        <span>{copy.zoomShort}: {params.zoom.toFixed(1)}</span>
-      </div>
-
-      <label className={clsx('mt-3 flex items-center gap-2 text-sm', !supportsMultiOutput ? 'text-text-muted' : 'text-text-secondary')}>
-        <input
-          type="checkbox"
-          className="h-4 w-4 rounded border-border"
-          checked={generateBestAngles}
-          disabled={!supportsMultiOutput}
-          onChange={(event) => {
-            const nextChecked = event.target.checked;
-            onGenerateBestAnglesChange(nextChecked);
-            if (nextChecked) {
-              onParamsChange(
-                sanitizeParams({
-                  rotation: 0,
-                  tilt: 0,
-                  zoom: params.zoom,
-                })
-              );
-            }
-          }}
-        />
-        {copy.bestAngles}
-      </label>
-    </div>
-  );
-}
 
 export default function AngleToolPage() {
   const { loading: authLoading, user } = useRequireAuth({ redirectIfLoggedOut: false });
@@ -1348,145 +1056,6 @@ export default function AngleToolPage() {
           </div>
         </div>
       ) : null}
-    </div>
-  );
-}
-
-function AngleImageLibraryModal({
-  open,
-  onClose,
-  onSelect,
-  copy,
-}: {
-  open: boolean;
-  onClose: () => void;
-  onSelect: (asset: LibraryAsset) => void;
-  copy: AngleCopy;
-}) {
-  const [activeSource, setActiveSource] = useState<'all' | 'upload' | 'generated'>('all');
-  const swrKey = open
-    ? activeSource === 'all'
-      ? '/api/user-assets?limit=60'
-      : `/api/user-assets?limit=60&source=${encodeURIComponent(activeSource)}`
-    : null;
-  const { data, error, isLoading } = useSWR<LibraryAssetsResponse>(swrKey, async (url: string) => {
-    const response = await authFetch(url);
-    const payload = (await response.json().catch(() => null)) as LibraryAssetsResponse | null;
-    if (!response.ok || !payload?.ok) {
-      throw new Error(copy.libraryError);
-    }
-    return payload;
-  });
-
-  const assets = data?.assets ?? [];
-
-  if (!open) return null;
-
-  const handleBackdropClick = (event: ReactMouseEvent<HTMLDivElement>) => {
-    if (event.target === event.currentTarget) {
-      onClose();
-    }
-  };
-
-  return (
-    <div
-      className="fixed inset-0 z-[10050] flex items-center justify-center bg-surface-on-media-dark-60 px-3 py-6 sm:px-6"
-      role="dialog"
-      aria-modal="true"
-      onMouseDown={handleBackdropClick}
-    >
-      <div className="max-h-[90vh] w-full max-w-3xl overflow-hidden rounded-modal border border-border bg-surface shadow-float">
-        <div className="flex flex-col gap-4 border-b border-border px-4 py-4 sm:flex-row sm:items-start sm:justify-between sm:px-6">
-          <div>
-            <h2 className="text-lg font-semibold text-text-primary">{copy.libraryChoose}</h2>
-            <p className="text-xs text-text-secondary">{copy.libraryBody}</p>
-          </div>
-          <Button
-            type="button"
-            variant="outline"
-            size="sm"
-            onClick={onClose}
-            className="self-start rounded-full border-border px-3 text-sm font-medium text-text-secondary hover:bg-bg sm:self-auto"
-          >
-            {copy.close}
-          </Button>
-        </div>
-        <div className="border-b border-border px-4 py-3 sm:px-6">
-          <div
-            role="tablist"
-            aria-label={copy.library}
-            className="flex w-full overflow-hidden rounded-full border border-border bg-surface-glass-70 text-xs font-semibold text-text-secondary"
-          >
-            {([
-              ['all', copy.tabs.all],
-              ['upload', copy.tabs.upload],
-              ['generated', copy.tabs.generated],
-            ] as const).map(([value, label]) => (
-              <Button
-                key={value}
-                type="button"
-                role="tab"
-                variant="ghost"
-                size="sm"
-                aria-selected={activeSource === value}
-                onClick={() => setActiveSource(value)}
-                className={clsx(
-                  'flex-1 rounded-none px-4 py-2',
-                  activeSource === value ? 'bg-brand text-on-brand hover:bg-brand' : 'text-text-secondary hover:bg-surface'
-                )}
-              >
-                {label}
-              </Button>
-            ))}
-          </div>
-        </div>
-        <div className="max-h-[70vh] overflow-y-auto px-4 py-4 sm:px-6">
-          {error ? (
-            <div className="rounded-card border border-state-warning/40 bg-state-warning/10 px-4 py-6 text-sm text-state-warning">
-              {error instanceof Error ? error.message : copy.libraryError}
-            </div>
-          ) : isLoading && !assets.length ? (
-            <div className="grid grid-gap-sm sm:grid-cols-2 lg:grid-cols-3">
-              {Array.from({ length: 6 }).map((_, index) => (
-                <div key={`angle-library-skeleton-${index}`} className="rounded-card border border-border bg-surface-glass-60 p-0" aria-hidden>
-                  <div className="relative aspect-square overflow-hidden rounded-t-card bg-placeholder">
-                    <div className="skeleton absolute inset-0" />
-                  </div>
-                  <div className="border-t border-border px-4 py-3">
-                    <div className="h-3 w-24 rounded-full bg-skeleton" />
-                  </div>
-                </div>
-              ))}
-            </div>
-          ) : assets.length === 0 ? (
-            <div className="rounded-card border border-dashed border-border px-4 py-6 text-center text-sm text-text-secondary">
-              {copy.libraryEmpty}
-            </div>
-          ) : (
-            <div className="grid grid-gap-sm sm:grid-cols-2 lg:grid-cols-3">
-              {assets.map((asset) => (
-                <button
-                  key={asset.id}
-                  type="button"
-                  onClick={() => onSelect(asset)}
-                  className="group block w-full overflow-hidden rounded-card border border-border bg-surface text-left shadow-card transition hover:border-text-primary focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-bg"
-                >
-                  <div className="relative aspect-square overflow-hidden rounded-t-card bg-placeholder">
-                    <img src={asset.url} alt="" className="h-full w-full object-cover" loading="lazy" referrerPolicy="no-referrer" />
-                    <div className="absolute inset-0 hidden items-center justify-center bg-surface-on-media-dark-40 text-sm font-semibold text-on-inverse group-hover:flex">
-                      {copy.useImage}
-                    </div>
-                  </div>
-                  <div className="min-w-0 space-y-1 border-t border-border px-4 py-3 text-xs text-text-secondary">
-                    <p className="truncate text-text-primary">{asset.url.split('/').pop() ?? copy.assetFallback}</p>
-                    {asset.createdAt ? <p className="text-text-muted">{new Date(asset.createdAt).toLocaleString()}</p> : null}
-                  </div>
-                </button>
-              ))}
-            </div>
-          )}
-        </div>
-      </div>
     </div>
   );
 }
